+++ R/combi_2.R
@@ -1,6 +1,6 @@
 #' @title GMDH COMBI auxiliar functions
 #'
-#' @description Performs auxiliar tasks to predict.mia
+#' @description Build a regression model performing GMDH Combinatorial with external criteria.
 #'
 #' @keywords internal
 #'
@@ -9,7 +9,10 @@ gmdh.combi_2 <- function(X, y, G = 2, x.test, y.test) {
   results <- vector(mode = "list", length = 2)
   names(results) <- c("results", "G")
   regressors <- fun.poly(X, G = G)
+  regressors <- fun.filter(regressors)
+
   x.test <- fun.poly(x.test, G = G)
+  x.test <- x.test[, colnames(regressors), drop = FALSE]
 
   combs <- do.call(expand.grid, rep(list(c(FALSE, TRUE)), ncol(regressors)))[-1, ]