+++ R/fun.svd_1.R
@@ -17,11 +17,13 @@ fun.svd_1 <- function(x, y) {
 
   Xsvd <- svd(x)
   D <- 1 / Xsvd$d
-  D[which(D < tol)] <- 0
+  D[D <= tol] <- 0
   C <- Xsvd$v %*% (crossprod(Xsvd$u, y) * D)
-  err <- (x %*% C) - y
   rownames(C) <- c("Ind", nombres)
+
+  err <- (x %*% C) - y
   CV <- sqrt(mean((err / (1 - rowSums(Xsvd$u * Xsvd$u)))^2))
+  CV <- round(CV, digits = 6)
 
   resultado$coef <- C
   resultado$CV <- CV