+++ R/gia.R
@@ -8,7 +8,7 @@
 #' @param y vector or matrix containing dependent variable in the model. \cr
 #' The data must not contain NAs
 #' @param prune an integer whose recommended minimum value is the number of initial regressors. \cr
-#'    The maximum value will depend on the available RAM. It is recommended to work with the maximum value, but it can be computationally very expensive. \cr
+#'    The maximum value will depend on the available RAM. \cr
 #'    Prune is the selected number of neurons from layer i to layer i+1. The resulting layer i+1 has prune(prune-1)/2 neurons; for example with prune=150, the resulting nerurons will be 11.175
 #' @param criteria GMDH external criteria. Values: \cr
 #'      \itemize{
@@ -24,29 +24,32 @@
 #' @return An object of class gia.
 #'
 #' @examples
-#' set.seed(123)
-#' x <- matrix(data = c(rnorm(1000)), ncol = 5, nrow = 200)
-#' colnames(x) <- c("a", "b", "c", "d", "e")
-
-#' y <- matrix(data = c(10 + x[, "a"] * x[, "e"]^3), ncol = 1)
-#' colnames(y) <- "y"
-#' x.test <- x[1:10, ]
-#' y.test <- y[1:10]
-#' x <- x[-c(1:10), ]
-#' y <- y[-c(1:10)]
 #'
-#' mod <- gmdh.gia(X = x, y = y, prune = 5, criteria = "PRESS")
+#' set.seed(123)
+#' x <- matrix(data = c(rnorm(500)), ncol = 4, nrow = 125)
+#' colnames(x) <- c("a", "b", "c", "d")
+
+#' y <- matrix(data = c(10 + x[, "a"] + x[, "d"]^2), ncol = 1)
+#' colnames(y) <- "y"
+#' x.test <- x[1:5, ]
+#' y.test <- y[1:5]
+#' x <- x[-c(1:5), ]
+#' y <- y[-c(1:5)]
+#'
+#' mod <- gmdh.gia(X = x, y = y, criteria = "PRESS")
 #' pred <- predict(mod, x.test)
 #' summary(sqrt((pred - y.test)^2))
 #'
 #' @references
 #' Bozdogan, H. and Haughton, D.M.A. (1998): "Information complexity criteria for regression models", Computational Statistics & Data Analysis, 28, pp. 51-76 <doi: 10.1016/S0167-9473(98)00025-5> \cr
 #'
-#' Farlow, S.J. (1981): "The GMDH algorithm of Ivakhnenko", The American Statistician, 35(4), pp. 210-215. <doi:10.2307/2683292> \cr
+#' Farlow, S.J. (1981): "The GMDH algorithm of Ivakhnenko", The American Statistician, 35(4), pp. 210-215. <doi: 10.2307/2683292> \cr
 #'
 #' Hild, Ch. R. and Bozdogan, H. (1995): "The use of information-based model selection criteria in the GMDH algorithm", Systems Analysis Modelling Simulation, 20(1-2), pp. 29-50 \cr
 #'
-#' Ivakhnenko A.G. (1968): "The Group Method of Data Handling - A Rival of the Method of Stochastic Approximation", Soviet Automatic Control, 13(3), pp. 43-55 \cr
+#' Ivakhnenko, A.G. (1968): "The Group Method of Data Handling - A Rival of the Method of Stochastic Approximation", Soviet Automatic Control, 13(3), pp. 43-55 \cr
+#'
+#' Müller, J.-A., Ivachnenko, A.G. and Lemke, F. (1998): "GMDH Algorithms for Complex Systems Modelling", Mathematical and Computer Modelling of Dynamical Systems, 4(4), pp. 275-316 <doi: 10.1080/13873959808837083> \cr
 #'
 #' Stepashko, V. Bulgakova, O. and Zosimov V. (2018): "Construction and Research of the Generalized Iterative GMDH Algorithm with Active Neurons", Advances in Intelligent Systems and Computing II, pp. 492-510 <doi:10.1007/978-3-319-70581-1_35>
 #'
@@ -56,10 +59,14 @@
 #'
 #' @export
 #'
-gmdh.gia <- function(X, y, prune = 40, criteria = c("PRESS", "test", "ICOMP"), x.test = NULL, y.test = NULL) {
+gmdh.gia <- function(X, y, prune = ncol(X), criteria = c("PRESS", "test", "ICOMP"), x.test = NULL, y.test = NULL) {
 
   if (is.matrix(X) != TRUE)
     stop("X must be a matrix")
+  if (is.null(colnames(X)) == TRUE)
+    stop("X matrix regressors must have names. Try colnames() function")
+  if(ncol(X) <= 2)
+    stop("GMDH GIA needs more than two regressors")
   if(any(is.na(X)) != FALSE)
     stop("X has NA values")
   if(any(is.na(y)) != FALSE)