+++ R/mia.R
@@ -8,7 +8,7 @@
 #' @param y vector or matrix containing dependent variable in the model. \cr
 #' The data must not contain NAs
 #' @param prune an integer whose recommended minimum value is the number of initial regressors. \cr
-#'    The maximum value will depend on the available RAM. It is recommended to work with the maximum value, but it can be computationally very expensive. \cr
+#'    The maximum value will depend on the available RAM. \cr
 #'    Prune is the selected number of neurons from layer i to layer i+1. The resulting layer i+1 has prune(prune-1)/2 neurons; for example with prune=150, the resulting nerurons will be 11.175
 #' @param criteria GMDH external criteria. Values: \cr
 #'      \itemize{
@@ -35,18 +35,20 @@
 #' x <- x[-c(1:10), ]
 #' y <- y[-c(1:10)]
 #'
-#' mod <- gmdh.mia(X = x, y = y, prune = 5, criteria = "PRESS")
+#' mod <- gmdh.mia(X = x, y = y, criteria = "PRESS")
 #' pred <- predict(mod, x.test)
 #' summary(sqrt((pred - y.test)^2))
 #'
 #' @references
 #' Bozdogan, H. and Haughton, D.M.A. (1998): "Information complexity criteria for regression models", Computational Statistics & Data Analysis, 28, pp. 51-76 <doi: 10.1016/S0167-9473(98)00025-5> \cr
 #'
-#' Farlow, S.J. (1981): "The GMDH algorithm of Ivakhnenko", The American Statistician, 35(4), pp. 210-215. <doi:10.2307/2683292> \cr
+#' Farlow, S.J. (1981): "The GMDH algorithm of Ivakhnenko", The American Statistician, 35(4), pp. 210-215. <doi: 10.2307/2683292> \cr
 #'
 #' Hild, Ch. R. and Bozdogan, H. (1995): "The use of information-based model selection criteria in the GMDH algorithm", Systems Analysis Modelling Simulation, 20(1-2), pp. 29-50 \cr
 #'
-#' Ivakhnenko A.G. (1968): "The Group Method of Data Handling - A Rival of the Method of Stochastic Approximation", Soviet Automatic Control, 13(3), pp. 43-55
+#' Ivakhnenko, A.G. (1968): "The Group Method of Data Handling - A Rival of the Method of Stochastic Approximation", Soviet Automatic Control, 13(3), pp. 43-55 \cr
+#'
+#' Müller, J.-A., Ivachnenko, A.G. and Lemke, F. (1998): "GMDH Algorithms for Complex Systems Modelling", Mathematical and Computer Modelling of Dynamical Systems, 4(4), pp. 275-316 <doi: 10.1080/13873959808837083>
 #'
 #' @importFrom stats na.omit na.fail predict
 #'
@@ -54,10 +56,14 @@
 #'
 #' @export
 #'
-gmdh.mia <- function(X, y, prune = 150, criteria = c("PRESS", "test", "ICOMP"), x.test = NULL, y.test = NULL) {
+gmdh.mia <- function(X, y, prune = ncol(X), criteria = c("PRESS", "test", "ICOMP"), x.test = NULL, y.test = NULL) {
 
   if (is.matrix(X) != TRUE)
     stop("X must be a matrix")
+  if (is.null(colnames(X)) == TRUE)
+    stop("X matrix regressors must have names. Try colnames() function")
+  if(ncol(X) <= 2)
+    stop("GMDH MIA needs more than two regressors")
   if(any(is.na(X)) != FALSE)
     stop("X has NA values")
   if(any(is.na(y)) != FALSE)