+++ R/predict.gia.R
@@ -9,18 +9,19 @@
 #' @return A matrix with predictions.
 #'
 #' @examples
-#' set.seed(123)
-#' x <- matrix(data = c(rnorm(1000)), ncol = 5, nrow = 200)
-#' colnames(x) <- c("a", "b", "c", "d", "e")
-
-#' y <- matrix(data = c(10 + x[, "a"] * x[, "e"]^3), ncol = 1)
-#' colnames(y) <- "y"
-#' x.test <- x[1:10, ]
-#' y.test <- y[1:10]
-#' x <- x[-c(1:10), ]
-#' y <- y[-c(1:10)]
 #'
-#' mod <- gmdh.gia(X = x, y = y, prune = 5, criteria = "PRESS")
+#' set.seed(123)
+#' x <- matrix(data = c(rnorm(500)), ncol = 4, nrow = 125)
+#' colnames(x) <- c("a", "b", "c", "d")
+
+#' y <- matrix(data = c(10 + x[, "a"] + x[, "d"]^2), ncol = 1)
+#' colnames(y) <- "y"
+#' x.test <- x[1:5, ]
+#' y.test <- y[1:5]
+#' x <- x[-c(1:5), ]
+#' y <- y[-c(1:5)]
+#'
+#' mod <- gmdh.gia(X = x, y = y, criteria = "PRESS")
 #' pred <- predict(mod, x.test)
 #' summary(sqrt((pred - y.test)^2))
 #'
@@ -43,6 +44,7 @@ predict.gia <- function(object, newdata, ...) {
            newdata <- fun.poly(inicial, G = 2),
            newdata <- fun.poly(newdata, G = 2)
     )
+
     for(j in 1:length(datos)) datos[[j]] <- predict.svd(object[[i]][[j]], newdata)
     newdata <- matrix(data = unlist(datos), ncol = length(datos))
     colnames(newdata) <- names(object[[i]])